backend-py/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.13.0
python-multipart==0.0.20
python-dotenv==1.0.1
